refactor(Page) : pinch -> longPress 이름 정정 수정
