+++ src/1_Page/Profile/ui/TrackContainer/index.js
@@ -1,13 +1,13 @@
 import React from "react";
 import Tracking from "../../../../2_Widget/TrackingImage";
 import STYLE from "./style";
-import usePinchEvent from "../../../../4_Shared/usePinchEvent";
+import useLongPressEvent from "../../../../4_Shared/useLongPressEvent";
 
-const Track = (props) => {
+const TrackContainer = (props) => {
   const { data, checkSetMode, author } = props;
   const { handleModifyMapOpen, setPinchedData } = props;
 
-  const pinchEvents = usePinchEvent(
+  const longPressEvents = useLongPressEvent(
     () => {
       handleModifyMapOpen();
       setPinchedData(data);
@@ -15,12 +15,13 @@ const Track = (props) => {
     null, // 꾸욱 누르기를 끝낼 때 동작 (필요 시 추가)
     1000 // 1초 꾸욱 누르기 감지
   );
+
   return (
-    <STYLE.TrackingContainer {...(author && pinchEvents)}>
+    <STYLE.TrackingContainer {...(author && longPressEvents)}>
       {checkSetMode && <STYLE.TrackingCheckbox />}
       <Tracking data={data} height={"100%"} />
     </STYLE.TrackingContainer>
   );
 };
 
-export default Track;
+export default TrackContainer;
