+++ src/2_Widget/ModalModifyTrackingImage/style.js
